+++ CHANGE_LOG.md
@@ -5,13 +5,13 @@
 - Add a flag to RulesContent signaling whether or not it contains input attestations
 
 # version 0.12.1-preview.17
-## JSONPath vulnerability fix
+## Complete Siop Validation
 **Type of change:** bug fix
 **Customer impact:** high
 - Full Siop Id Token Validation
 
 # version 0.12.1-preview.16
-## JSONPath vulnerability fix
+## Allow pre-validation of a siop token for issuance
 **Type of change:** bug fix
 **Customer impact:** low
 - Allow for a siop token to be validated before calling Validator.validate